--- hands-on/listings-workshop/throttlepercentage.json
@@ -1,98 +1,0 @@
-[
-    {
-        "bin(1h)": "2025-06-01 23:00:00.000",
-        "percentagePatchThrottles429": "15.4519"
-    },
-    {
-        "bin(1h)": "2025-06-01 22:00:00.000",
-        "percentagePatchThrottles429": "16.7202"
-    },
-    {
-        "bin(1h)": "2025-06-01 21:00:00.000",
-        "percentagePatchThrottles429": "16.0848"
-    },
-    {
-        "bin(1h)": "2025-06-01 20:00:00.000",
-        "percentagePatchThrottles429": "11.7766"
-    },
-    {
-        "bin(1h)": "2025-06-01 19:00:00.000",
-        "percentagePatchThrottles429": "26.3193"
-    },
-    {
-        "bin(1h)": "2025-06-01 18:00:00.000",
-        "percentagePatchThrottles429": "26.4372"
-    },
-    {
-        "bin(1h)": "2025-06-01 17:00:00.000",
-        "percentagePatchThrottles429": "16.614"
-    },
-    {
-        "bin(1h)": "2025-06-01 16:00:00.000",
-        "percentagePatchThrottles429": "17.8858"
-    },
-    {
-        "bin(1h)": "2025-06-01 15:00:00.000",
-        "percentagePatchThrottles429": "25.6745"
-    },
-    {
-        "bin(1h)": "2025-06-01 14:00:00.000",
-        "percentagePatchThrottles429": "23.1226"
-    },
-    {
-        "bin(1h)": "2025-06-01 13:00:00.000",
-        "percentagePatchThrottles429": "20.6362"
-    },
-    {
-        "bin(1h)": "2025-06-01 12:00:00.000",
-        "percentagePatchThrottles429": "14.9395"
-    },
-    {
-        "bin(1h)": "2025-06-01 11:00:00.000",
-        "percentagePatchThrottles429": "14.618"
-    },
-    {
-        "bin(1h)": "2025-06-01 10:00:00.000",
-        "percentagePatchThrottles429": "16.0751"
-    },
-    {
-        "bin(1h)": "2025-06-01 09:00:00.000",
-        "percentagePatchThrottles429": "17.4411"
-    },
-    {
-        "bin(1h)": "2025-06-01 08:00:00.000",
-        "percentagePatchThrottles429": "30.6383"
-    },
-    {
-        "bin(1h)": "2025-06-01 07:00:00.000",
-        "percentagePatchThrottles429": "17.4633"
-    },
-    {
-        "bin(1h)": "2025-06-01 06:00:00.000",
-        "percentagePatchThrottles429": "26.8644"
-    },
-    {
-        "bin(1h)": "2025-06-01 05:00:00.000",
-        "percentagePatchThrottles429": "27.9257"
-    },
-    {
-        "bin(1h)": "2025-06-01 04:00:00.000",
-        "percentagePatchThrottles429": "20.7373"
-    },
-    {
-        "bin(1h)": "2025-06-01 03:00:00.000",
-        "percentagePatchThrottles429": "32.6645"
-    },
-    {
-        "bin(1h)": "2025-06-01 02:00:00.000",
-        "percentagePatchThrottles429": "30.2524"
-    },
-    {
-        "bin(1h)": "2025-06-01 01:00:00.000",
-        "percentagePatchThrottles429": "38.0271"
-    },
-    {
-        "bin(1h)": "2025-06-01 00:00:00.000",
-        "percentagePatchThrottles429": "14.5735"
-    }
-]
