+++ labs/README.md
@@ -1,0 +1,33 @@
+# 🧪 SP-API Labs
+
+Welcome to the **SP-API Labs**!  
+This directory contains hands-on resources to help you learn and master the **Selling Partner API (SP-API)** through code samples, guided tutorials, and full workshop challenges.
+
+---
+
+## 📂 Structure
+
+- **`demos/`**  
+  Short, focused examples that show how to use the SP-API SDK for specific operations.  
+  - Ideal if you want a quick code reference or “hello world”–style snippet.  
+  - Example: calling `getOrders` with the SDK in Python.
+
+- **`tutorials/`**  
+  Step-by-step Jupyter notebooks that guide you through individual workflows.  
+  - Ideal if you want to learn and test one API domain in detail.  
+  - Example: querying with the Data Kiosk API.
+
+- **`workshops/`**  
+  End-to-end labs that combine multiple APIs to solve a real-world problem.  
+  - Ideal for deeper learning or group workshops.  
+  - Example: diagnosing listing errors, or building a fulfillment solution with Buy Shipping + MLI.
+
+---
+
+## 📚 Quick Access Table
+
+| Category    | Folder        | Description                                                                 | Example                                                |
+|-------------|---------------|-----------------------------------------------------------------------------|--------------------------------------------------------|
+| 💡 Demos    | `demos/`      | Short code snippets using the SP-API SDK                                    | GetOrders with the Python SDK                          |
+| 📖 Tutorials | `tutorials/` | Guided Jupyter notebooks focused on a single API workflow                   | Querying SKU economics with Data Kiosk                 |
+| 🏋️ Workshops | `workshops/` | Multi-step challenges combining multiple APIs to solve real-world problems | Listing error diagnosis, Fulfillment workflows with MLI |
