+++ labs/demos/README.md
@@ -1,0 +1,20 @@
+# 💡 SP-API Demos
+
+This folder contains **short, focused code examples** that demonstrate how to use the SP-API SDK for specific operations.  
+Think of these as quick-start "hello world" references for developers.
+
+---
+
+## 🧑‍💻 What Are Demos?
+
+- **Minimal** → only show the API call and basic setup.  
+- **SDK-first** → focus on the official SP-API SDK usage.  
+- **Language-agnostic** → may include examples in Java, Python, JavaScript, etc.  
+
+---
+## 📚 Available Demos
+
+
+| Demo                       | File/Folder             | Language | APIs Used        | Description                                                                 |
+|----------------------------|-------------------------|----------|------------------|-----------------------------------------------------------------------------|
+| AWD Inbound Workflow Demo  | `awd/AWDDemoSolution.java`  | Java     | AWD (2024-05-09) | End-to-end AWD flow: check eligibility → create inbound → get inbound → confirm inbound |
