+++ labs/workshops/README.md
@@ -1,0 +1,33 @@
+# 🏋️ SP-API Workshops
+
+Welcome to the **SP-API Workshops** folder!  
+Here you’ll find hands-on challenges designed to simulate real-world seller problems and give you practical experience with the **Selling Partner API**.
+
+Each workshop:
+- Provides a **scenario** based on real seller issues.
+- Walks you through **multiple SP-API workflows**.
+- Includes **sample payloads, hints, and mock endpoints** for safe experimentation.
+- Can be run in **AWS SageMaker (recommended)** or locally via Jupyter.
+
+---
+
+## 🚀 How to Get Started
+
+1. **SageMaker (Workshop Mode)**  
+   - Open the launch URL shared during your session.  
+   - Log in with provided credentials.  
+   - Go to **Amazon SageMaker → Notebook Instances**.  
+   - Open Jupyter and navigate to `labs/workshops/`.
+
+2. **Local Jupyter (Self-Paced Mode)**  
+   ```bash
+   git clone https://github.com/amzn/sp-api-sample-solutions.git
+   cd labs/workshops
+   pip install -r ../env/requirements.txt
+   jupyter lab```
+
+## 📚 Available Challenges
+
+| Challenge             | Folder              | Focus APIs                         | Scenario                                                       |
+|-----------------------|---------------------|------------------------------------|----------------------------------------------------------------|
+| 🧙 ListingsWizard     | `listings-wizard/`   | Listings, Pricing, Data Kiosk      | Optimize listings conversion, pricing, and refund rates        |
